Bot/requirements.txt
py-cord
pythonping 
requests
psutil
simpcalci
https://github.com/ytdl-org/youtube-dl/archive/refs/heads/master.zip
yt-dlp
youtube-search-python
ffmpeg
PyNaCl
fake_useragent
opencc-python-reimplemented
aiofiles
pixivpy3
beautifulsoup4
bs4
certifi
chardet
idna
pillow
pytesseract
soupsieve
urllib3
pycord-ext-ipc
Quart-Discord-any[pycord]
spotipy
ytmusicapi
rich
pycord-i18n
GitPython